+++ src/commands/cd.js
@@ -99,7 +99,7 @@ const cd = (dir, ...args) => {
     } else if (err.code === "ENOENT") {
       // For some reason, there are rare cases where the checks think the directory exists,
       // but when trying to change into it, it throws an error.
-      // This usually happens when using the BubbleOS executable, where a folder called
+      // This usually happens when using the BubbleOS executable, where a directory called
       // "C:\snapshot" is visible on Windows (through 'ls' in BubbleOS), but when trying to
       // change into it, it throws an error.