+++ src/commands/stat.js
@@ -130,7 +130,7 @@ const stat = (path = process.cwd(), ...args) => {
     } else if (err.code === "ENOENT") {
       // For some reason, there are rare cases where the checks think the path exists,
       // but when trying to check stats, it throws an error.
-      // This usually happens when using the BubbleOS executable, where a folder called
+      // This usually happens when using the BubbleOS executable, where a directory called
       // "C:\snapshot" is visible on Windows (through 'ls' in BubbleOS), but when trying to
       // get stats, it throws an error.