+++ CHANGELOG.md
@@ -11,6 +11,8 @@ _No added features._
 ### Changed/Fixed Features
 
 - Changed the `stat` command to automatically retrieve the stats of the current directory if no path is specified.
+- Changed all references from "folder" to "directory" to help create a consistency of terminology usage.
+- Changed the `sysinfo` CPU information so that the speed will not be outputted if it is zero, and the CPU information will not be shown if BubbleOS is unable to get CPU information.
 - Fixed an issue where reading an empty directory with `dirtree` would not show a message saying the directory was empty.
 
 ### Removed Features